refactor(pdf-service): move removeDecimalsFromCurrencyString to /app/util/strings
