+++ app/domains/prozesskostenhilfe/services/pdf/__test__/util.test.ts
@@ -1,7 +1,4 @@
-import {
-  getTotalMonthlyFinancialEntries,
-  removeDecimalsFromCurrencyString,
-} from "../util";
+import { getTotalMonthlyFinancialEntries } from "../util";
 
 describe("utility functions", () => {
   describe("getTotalMonthlyFinancialEntries", () => {
@@ -46,28 +43,6 @@ describe("utility functions", () => {
           },
         ]),
       ).toBe("1.120,00");
-    });
-  });
-
-  describe("removeDecimalsFromCurrencyString", () => {
-    it("should remove decimals from a currency string", () => {
-      const result = removeDecimalsFromCurrencyString("1.234,56");
-      expect(result).toBe("1.234");
-    });
-
-    it("should return the same string if there are no decimals", () => {
-      const result = removeDecimalsFromCurrencyString("1.234");
-      expect(result).toBe("1.234");
-    });
-
-    it("should handle an empty string", () => {
-      const result = removeDecimalsFromCurrencyString("");
-      expect(result).toBe("");
-    });
-
-    it("should handle strings without currency values", () => {
-      const result = removeDecimalsFromCurrencyString("Hello World");
-      expect(result).toBe("Hello World");
     });
   });
 });
