+++ app/domains/prozesskostenhilfe/services/pdf/util.ts
@@ -1,7 +1,5 @@
 import type { FinancialEntry } from "~/domains/shared/formular/finanzielleAngaben/context";
 
-export const nettoString = "netto";
-
 export const getTotalMonthlyFinancialEntries = (
   financialEntries: FinancialEntry[],
 ) =>
@@ -26,10 +24,3 @@ export const getTotalMonthlyFinancialEntries = (
       minimumFractionDigits: 2,
       maximumFractionDigits: 2,
     });
-
-export function removeDecimalsFromCurrencyString(
-  currencyString: string | undefined,
-) {
-  if (currencyString === undefined) return;
-  return currencyString.replace(/,\d{2}/g, "");
-}
